requirements.txt
Flask==2.0.1
Pillow==9.0.0
gunicorn==20.1.0 